Some telemetry improvements...

+++ internal/client/ollama/contract.go
@@ -93,12 +93,12 @@ type GenerateResponse struct {
 	CreatedAt          time.Time `json:"created_at"`
 	Response           string    `json:"response"`
 	Done               bool      `json:"done"`
-	Context            []int     `json:"context"`
+	Context            []int64   `json:"context"`
 	TotalDuration      int64     `json:"total_duration"`
-	LoadDuration       int       `json:"load_duration"`
-	PromptEvalCount    int       `json:"prompt_eval_count"`
-	PromptEvalDuration int       `json:"prompt_eval_duration"`
-	EvalCount          int       `json:"eval_count"`
+	LoadDuration       int64     `json:"load_duration"`
+	PromptEvalCount    int64     `json:"prompt_eval_count"`
+	PromptEvalDuration int64     `json:"prompt_eval_duration"`
+	EvalCount          int64     `json:"eval_count"`
 	EvalDuration       int64     `json:"eval_duration"`
 }
